Fix model downloads for ignitionrobotics.org URIs (#318)
The FuelClient library was only looking for models in ~/.ignition/fuel/fuel.gazebosim.org regardless of which server models were downloaded from when checking if a model is cached. This caused models that used the deprecated ignitionrobotics.org URI to be redownloaded every time. This PR fixes it by adding ignitionrobotics.org as an additional server to check.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>