Update package deps for harmonic (#326)
Now that msgs and transport are recently bumped (to msgs10 and transport13) in harmonic, we need to bump all packages that depend on them. In this case, harmonic should now use fuel-tools9.

I noticed that the package dependency versions were bumped in main in 83077ca. Here I'm just updating all of them to correspond to the correct gz package versions that we currently use in harmonic

Signed-off-by: Ian Chen <ichen@openrobotics.org>